requirements.txt
agno
python-dotenv
openai
psycopg[binary]
pgvector
sqlalchemy
fastapi
uvicorn